pyproject.toml
[project]
name = "llm-service"
version = "0.1.0"
description = "gRPC service to chat with LLMs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.2",
    "asyncio>=4.0.0",
    "asyncpg>=0.30.0",
    "cohere>=5.20.0",
    "fastapi>=0.122.0",
    "grpcio>=1.75.1",
    "grpcio-tools>=1.75.1",
    "langchain>=1.0.8",
    "langchain-core>=1.0.7",
    "langchain-huggingface>=1.0.1",
    "langchain-ollama>=1.0.0",
    "langchain-qdrant>=1.1.0",
    "langgraph>=1.0.3",
    "langgraph-checkpoint-postgres>=3.0.1",
    "llmrouter-lib>=0.3.1",
    "mypy-protobuf>=3.7.0",
    "numpy>=2.3.3",
    "protobuf>=6.32.1",
    "psycopg[binary,pool]>=3.2.13",
    "pydantic>=2.12.0",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.15.1",
    "redis[hiredis]>=7.1.0",
    "routellm[eval,serve]>=0.2.0",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=5.1.2",
    "tiktoken>=0.12.0",
    "transformers>=4.57.1",
    "typing>=3.10.0.0",
    "valkey-glide>=2.2.3",
    "xgboost>=3.1.2",
]

[tool.basedpyright]
typeCheckingMode = "standard"  # Switching from "strict" to "standard"
venvPath = "."
venv = ".venv"
reportUnknownMemberType = false
reportUnknownVariableType = false
reportUnknownArgumentType = false

[dependency-groups]
dev = [
    "mypy-protobuf>=3.7.0",
]
